feat: order-history and  inscrese decrease quantity in detail product

+++ src/main/java/com/example/laptopshop_project/domain/OrderDetail.java
@@ -16,7 +16,7 @@ public class OrderDetail {
     private long id;
 
     private double price;
-    private int quantity;
+    private long quantity;
 
     @ManyToOne
     @JoinColumn(name = "product_id")
